chore: initial setup for ScreenStackHeaderSubview
