+++ ios/RNSScreenStackHeaderConfig.mm
@@ -36,16 +36,6 @@
 @end
 #endif
 
-@interface RNSScreenStackHeaderSubview : UIView
-
-@property (nonatomic, weak) RCTBridge *bridge;
-@property (nonatomic, weak) UIView *reactSuperview;
-@property (nonatomic) RNSScreenStackHeaderSubviewType type;
-
-- (instancetype)initWithBridge:(RCTBridge *)bridge;
-
-@end
-
 @implementation RNSScreenStackHeaderSubview
 
 - (instancetype)initWithBridge:(RCTBridge *)bridge
